+++ apps/native-component-list/src/screens/Video/VideoScreen.tsx
@@ -49,6 +49,7 @@ const eventsToListen: (keyof VideoPlayerEvents)[] = [
   'volumeChange',
   'playToEnd',
   'sourceChange',
+  'progressUpdate',
 ];
 if (Platform.OS === 'android') {
   videoLabels.push('Tears of Steel (DRM protected)');
@@ -80,6 +81,7 @@ export default function VideoScreen() {
     player.staysActiveInBackground = staysActiveInBackground;
     player.showNowPlayingNotification = true;
     player.allowsExternalPlayback = true;
+    player.progressUpdateInterval = 0.25;
     player.play();
   });